requirements.txt
openai==1.12.0
aiohttp==3.9.1
pydub==0.25.1
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
python-dotenv
pydantic
